Default input changes
Default inputs will no longer demand that the session is in a specific gamemode. They will all, instead, require that the survivor not be speaking any other scene.

This only applies to the default inputs for L4D1 and L4D2 concepts. Extra concepts such as PlayerYellRun will not be affected by this.