Population_Versus Update
Removes the fallen survivors (parachutists) from the versus version of Swamp Fever to be consistent with the other versus maps. In TLS, fallen survivors were added to different campaign levels, but each was removed from the versus version of the map for game balance. Since c3m2 already has generous supplies for survivors, there is no need for parachutists to spawn in the versus map.