Rainy Infected Variant Update
Updates all Rainy infected variants to use correct phong, detail scale, and color gradient. The Rainy variants will now be consistent with the Normal and Swamp infected variants.

Also  updates the unused "cif_tanktop_jeans_swamp.vmt" to be consistent with other swamp variants. The "cif_tanktop_jeans_swamp.vtf" file is missing, because it was never completed. However, if it is ever created it will be fully functional.