Added Chapter 1 cases
In the event that a chapter 1 intro does not play via the talker, the vocalizations will then resume after 20 seconds (Longest time in vanilla is The Sacrifice, between 15 and 16 seconds). If it does play, then vocalizations will resume once the conversation ends. 

This does not cover saferoom intros due to inaccurate results.