NoIntro Rules
Added a saferoom/non-chapter 1 variant of the nointro idle rule. In theory, should no one have the first line of an intro, that isn't the first map of the campaign, this rule would trigger for any survivor and would allow vocalizations after 5 or so seconds... which should also work with the rule weight system.

This is NOT fully functional yet, an internal tweak is needed to fix this, but I'm uploading the implementation now in preparation. Please do not report this as not working.