Block the gap left unclipped using `func_clip_vphysics`
Typically the roofs of the cola event's buildings have "vphysics clips" to prevent the cola being thrown out of bounds. Unfortunately, a narrow gap narrow was left unclipped.

https://cdn.discordapp.com/attachments/892506389379944449/1145359898239049859/l4d2_colatime.mp4

Since the clip entities too exist in survival, this commit brings them and the other past changes to survival:
- The nav blocker under the destroyed tanker's head.
- The +2 seconds of I/O delay to `tank_destruction_relay` entity, could be for Whittaker's VOs?
- The edited "info_map_parameters" settings.

Besides the above changes, the gap is occupied by a leftover trigger entity that is meant to delete the cola but doesn't actually does so, so this commit converts it to a `func_clip_vphysics` entity. The following video demonstrates the new change from this commit:
https://cdn.discordapp.com/attachments/1102216572170407986/1147832004528844800/l4d2_colatimersoutta.mp4