Disabled mip-maps on illum masks for the trucks
When viewed at a distance, the glow effect would bleed into the front lights of the vehicles, likely because the UV islands for the windshields and front lights are too close to each other. Hopefully this doesn't result in the glow effect looking aliased, but from what I could observe, it looked okay. Please confirm this for me.