Updated Dead Air 3 nav and lump 18 brushes
When Dead Air 3 was recompiled for L4D2, an extra brush clip that blocked Survivors from the airport rooftop also erroneously blocked the hallway next to the end safe room. The pipe room also has an Infected ladder up to a pipe, but the pipe was clipped off from Infected players. This is a very careful, minimalist edit of lump 18 brushes to open the hallway to Survivors (Infected already had access) and the pipe to Infected-only. The nav is updated to remove PLAYER_CLIP from the hallway, and allow 3 drop-down points from the pipe in case a player-Tank gets transferred to a bot (otherwise the bot might rub the wall and die). Note that "anv_mapfixes" contains a new make_clip called "lmpedit" to re-block the aforementioned airport roof. More info on lump edits @ https://github.com/Tsuey/L4D2-BSP-Lump-Editing